% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp21082R.R
\name{dep}
\alias{dep}
\title{Generate random sample from a two-component normal mixture using R}
\usage{
dep(mu, sigma, p)
}
\arguments{
\item{mu}{sample mean}

\item{sigma}{sample standard deviation}

\item{p}{ratio of two distributions}
}
\value{
random sample from a two-component normal mixture
}
\description{
Generate random sample from a two-component normal mixture using R
}
\examples{
\dontrun{
p=0.3 
mu=c(0,1) 
sigma =c(1,0.3)
dep(mu,sigma,p)
}
}
