% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp21082R.R
\name{floyd}
\alias{floyd}
\title{Find the shortest path between two vertex using R.}
\usage{
floyd(A)
}
\arguments{
\item{A}{primitive distance matrix}
}
\value{
the shortest distance between i to j
}
\description{
Find the shortest path between two vertex using R.
}
\examples{
\dontrun{
a=matrix(0,7,7)
a[1,2]=3;a[1,3]=5;
a[2,3]=1;a[2,4]=5;a[2,5]=8;
a[3,4]=7;a[3,5]=4;a[3,6]=10;
a[4,5]=3;a[4,7]=6;
a[5,6]=1;a[5,7]=2;
a[6,7]=2
b=a+t(a)
b[b==0]=Inf
floyd(b)
}
}
