% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gibbs_sampleC}
\alias{gibbs_sampleC}
\title{A Gibbs sampler using Rcpp}
\usage{
gibbs_sampleC(N, a, b, n, thin)
}
\arguments{
\item{N}{the length of chain}

\item{a}{parameter}

\item{b}{parameter}

\item{n}{the number of samples}

\item{thin}{the number of between-sample random numbers}
}
\value{
a random sample of size \code{n}
}
\description{
A Gibbs sampler using Rcpp
}
\examples{
\dontrun{
gibbsC <- gibbs_sampleC(2000,1,4,50,10)
qqplot(gibbsC[,1],gibbsR[,1],pch=16,cex=0.7,main="compare x") 
abline(a=0,b=1,col = "blue")
}
}
