% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp21082R.R
\name{kerden}
\alias{kerden}
\title{Using Gauss function as kernel density estimation using R}
\usage{
kerden(x, h)
}
\arguments{
\item{x}{samples}

\item{h}{bandwidth}
}
\value{
kernel density estimation
}
\description{
Using Gauss function as kernel density estimation for random sample of \code{dep} using R
}
\examples{
\dontrun{
p=0.3 
mu=c(0,1) 
sigma=c(1,0.3)
data=c(dep(mu,sigma,p))
plot(density(data),ylim=c(-0.1,5),main="",xlab="",col=1,lwd=2)
lines(kerden(data,0.005),col=2,lwd=1)
}
}
