\name{StatComp21082-package}
\alias{StatComp21082-package}
\alias{StatComp21082}
\docType{package}
\title{
  This package provides function for graph and  kernel density estimation
}
\description{
  Find the shortest path between two vertex using R;
  Using Gauss function as kernel density estimation using R
}
\author{
Lin YuYing
Maintainer: Lin YuYing <linyuying@mail.ustc.edu.cn>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
